test: divide tests into separate files
